requirements.txt
MouseInfo==0.1.3
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
